Fix format_push_string lint
This change fixes a lint introduced with the release of Rust 1.62 where
an unnecessary heap allocation is made.  This can be avoided by using a
series of push_str statements.  This is an alternative to using the
write macro, as suggested, as it would require the signature to change
in order to propagate a Error required by the general std::fmt::Write
trait.

See
https://rust-lang.github.io/rust-clippy/master/index.html#format_push_string
for more details.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>