Update changelog after pruning
This change updates the changelog after pruning by taking into account
pruning occurring in the middle of a chain of state roots.  The
changelog is updated such that the parent and successor nodes are
properly re-linked with the removal of the intermediate node.  It also
limits the pruning of changelog additions to those candidates that were
deleted.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>